+++ types/poisson-disk-sampling/src/implementations/variable-density.d.ts
@@ -1,0 +1,26 @@
+declare class VariableDensityPDS {
+    constructor(
+        options: {
+            shape: number[];
+            minDistance: number;
+            maxDistance?: number;
+            tries?: number;
+            distanceFunction: (...params: number[]) => number;
+            bias?: number;
+        },
+        rng?: ((...params: number[]) => number) | null,
+    );
+
+    addRandomPoint(): number[];
+    addPoint(point: number[]): number[] | null;
+
+    protected directAddPoint(point: number[]): number[];
+    protected inNeighbourhood(point: number[]): boolean;
+
+    next(): number[] | null;
+    fill(): number[][];
+    getAllPoints(): number[][];
+    getAllPointsWithDistance(): number[][];
+    reset(): void;
+}
+export = VariableDensityPDS;
