+++ types/poisson-disk-sampling/src/neighbourhood.d.ts
@@ -1,0 +1,2 @@
+export function getNeighbourhoodMemoized(dimensionNumber: number): number[];
+export function getNeighbourhood(dimensionNumber: number): number[];
