+++ types/poisson-disk-sampling/src/sphere-random.d.ts
@@ -1,0 +1,1 @@
+export function sampleSphere(d: number, rng: (...params: number[]) => number): number[];
