+++ types/poisson-disk-sampling/src/tiny-ndarray.d.ts
@@ -1,0 +1,1 @@
+export const tinyNDArray: any;
